taskResultPublisher module is introduced
