+++ tasks/webpack-build-task/index.ts
@@ -1,16 +1,6 @@
 import tl = require("vsts-task-lib/task");
 import { compile, IWebpackCompilationResult } from "./webpackCompiler";
-import { createWebpackResultMarkdownFile } from "./summarySectionBuilder";
-
-const convertMessageToSingleLine = (message: string): string => {
-    const messageParts = message.split("\n");
-
-    for (let index = 0; index < messageParts.length; index++) {
-        messageParts[index] = messageParts[index].trim();
-    }
-
-    return messageParts.join("; ");
-};
+import publishTaskResult from "./taskResultPublisher";
 
 async function run(): Promise<void> {
     let taskDisplayName = tl.getVariable("task.displayname");
@@ -29,10 +19,6 @@ async function run(): Promise<void> {
         const treatErrorsAs = tl.getInput("treatErrorsAs", true);
         const treatWarningsAs = tl.getInput("treatWarningsAs", true);
 
-        const errors = "errors";
-        const warnings = "warnings";
-        const info = "info";
-
         if (!workingFolder) {
             workingFolder = __dirname;
         }
@@ -49,58 +35,8 @@ async function run(): Promise<void> {
         process.chdir(workingFolder);
 
         const result = compile(workingFolder, webpackCliLocation, webpackCliArguments, webpackStatsJsLocation);
+        publishTaskResult(taskDisplayName, result, treatErrorsAs, treatWarningsAs, workingFolder, webpackStatsJsLocation);
 
-        const errorsArray: string[] = result.errors;
-        const warningsArray: string[] = result.warnings;
-
-        const hasErrors = errorsArray.length > 0;
-        const hasWarnings = warningsArray.length > 0;
-
-        if ((hasErrors && treatErrorsAs === errors) || (hasWarnings && treatWarningsAs === errors)) {
-            tl.setResult(tl.TaskResult.Failed, `${taskDisplayName} failed`);
-        } else if ((hasErrors && treatErrorsAs === warnings) || (hasWarnings && treatWarningsAs === warnings)) {
-            tl.warning(`${taskDisplayName} partially succeeded`);
-        } else {
-            console.log(`${taskDisplayName} succeeded`);
-        }
-
-        if (hasErrors && treatErrorsAs !== info) {
-            for (let errorItem of errorsArray) {
-                errorItem = `${taskDisplayName}: ${convertMessageToSingleLine(errorItem)}`;
-
-                if (treatErrorsAs === errors) {
-                    tl.error(errorItem);
-                } else if (treatErrorsAs === warnings) {
-                    tl.warning(errorItem);
-                }
-            }
-        }
-
-        if (hasWarnings && treatWarningsAs !== info) {
-            for (let warningItem of warningsArray) {
-                warningItem = `${taskDisplayName}: ${convertMessageToSingleLine(warningItem)}`;
-
-                if (treatWarningsAs === errors) {
-                    tl.error(warningItem);
-                } else if (treatWarningsAs === warnings) {
-                    tl.warning(warningItem);
-                }
-            }
-        }
-
-        const taskFailed = (hasErrors && treatErrorsAs === errors) || (hasWarnings && treatWarningsAs === errors);
-        const taskPartiallySucceeded = !taskFailed && ((hasErrors && treatErrorsAs === warnings) || (hasWarnings && treatWarningsAs === warnings));
-
-        if (taskPartiallySucceeded) {
-            console.log("##vso[task.complete result=SucceededWithIssues;]DONE");
-        }
-
-        try {
-            createWebpackResultMarkdownFile(workingFolder, taskDisplayName, result, webpackStatsJsLocation);
-        } catch (error) {
-            console.log("Couldn't create webpack result markdown file");
-            console.log(error);
-        }
     } catch (err) {
         tl.setResult(tl.TaskResult.Failed, `${taskDisplayName} failed`);
         tl.error(err);
